# apis/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/collections/{collection}": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Возвращает информацию о коллекции по ее имени или ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collections"
                ],
                "summary": "Просмотреть коллекцию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя или ID коллекции",
                        "name": "collection",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apis.Collection"
                        }
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Удаляет коллекцию по ее имени или ID",
                "tags": [
                    "Collections"
                ],
                "summary": "Удалить коллекцию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя или ID коллекции",
                        "name": "collection",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not found.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Обновляет информацию о коллекции по ее имени или ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collections"
                ],
                "summary": "Обновить коллекцию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя или ID коллекции",
                        "name": "collection",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления коллекции",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.CollectionCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apis.Collection"
                        }
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/collections/{collection}/records": {
            "get": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "description": "Возвращает список записей из указанной коллекции",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Record"
                ],
                "summary": "Получение списка записей",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор коллекции",
                        "name": "collection",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Получение списка записей успешно"
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "description": "Создает новую запись в указанной коллекции",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Record"
                ],
                "summary": "Создание записи",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор коллекции",
                        "name": "collection",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для создания записи",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.CreateRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Создание записи успешно"
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/collections/{collection}/records/{id}": {
            "get": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "description": "Возвращает информацию о указанной записи из указанной коллекции",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Record"
                ],
                "summary": "Просмотр записи",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор коллекции",
                        "name": "collection",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Идентификатор записи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Просмотр записи успешен"
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "description": "Удаляет указанную запись из указанной коллекции",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Record"
                ],
                "summary": "Удаление записи",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор коллекции",
                        "name": "collection",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Идентификатор записи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Удаление записи успешно"
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "Auth": []
                    }
                ],
                "description": "Обновляет информацию о указанной записи в указанной коллекции",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Record"
                ],
                "summary": "Обновление записи",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор коллекции",
                        "name": "collection",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Идентификатор записи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления записи",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.CreateRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновление записи успешно"
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apis.Collection": {
            "type": "object",
            "properties": {
                "createRule": {
                    "type": "string"
                },
                "created": {
                    "type": "object"
                },
                "deleteRule": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "indexes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "listRule": {
                    "description": "rules",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "options": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "schema": {
                    "type": "object"
                },
                "system": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                },
                "updateRule": {
                    "type": "string"
                },
                "updated": {
                    "type": "object"
                },
                "viewRule": {
                    "type": "string"
                }
            }
        },
        "apis.CollectionCreateRequest": {
            "type": "object",
            "properties": {
                "createRule": {
                    "type": "string"
                },
                "deleteRule": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "indexes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "listRule": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "options": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "schema": {
                    "type": "object"
                },
                "system": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                },
                "updateRule": {
                    "type": "string"
                },
                "viewRule": {
                    "type": "string"
                }
            }
        },
...
        "apis.CreateRecordRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "emailVisibility": {
                    "type": "boolean"
                },
                "id": {
                    "description": "base model fields",
                    "type": "string"
                },
                "oldPassword": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "passwordConfirm": {
                    "type": "string"
                },
                "username": {
                    "description": "auth collection fields\n---",
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        },
...
    }
}